Create strata-loader project
Move FpML code from strata-finance
