--- modules/finance/src/main/java/com/opengamma/strata/finance/loader/package-info.java
@@ -1,12 +1,0 @@
-/**
- * Copyright (C) 2015 - present by OpenGamma Inc. and the OpenGamma group of companies
- * 
- * Please see distribution for license.
- */
-
-/**
- * Loader that can convert files to financial instruments.
- * <p>
- * The main parser is for FpML. See {@link com.opengamma.strata.finance.loader.FpmlDocumentParser}.
- */
-package com.opengamma.strata.finance.loader;
