+++ modules/loader/README.md
@@ -1,0 +1,23 @@
+Strata-Loader
+-------------
+This directory contains the `strata-loader` module.
+
+### Overview
+
+This module provides tools that load data into the system from various data formats:
+
+* FpML
+* CSV files
+
+
+### Source code
+
+This module is released as Open Source Software using the
+[Apache v2.0 license](http://www.apache.org/licenses/LICENSE-2.0.html).  
+Commercial support is [available](http://www.opengamma.com/) from the authors.
+
+Code in this directory is not currently released.
+Classes and Methods may change at any time.
+Once released it will be maintained with backwards compatibility in mind.
+
+[![OpenGamma](http://developers.opengamma.com/res/display/default/chrome/masthead_logo.png "OpenGamma")](http://www.opengamma.com)
