+++ modules/loader/src/main/java/com/opengamma/strata/loader/fpml/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * Copyright (C) 2015 - present by OpenGamma Inc. and the OpenGamma group of companies
+ * 
+ * Please see distribution for license.
+ */
+
+/**
+ * Loader that can convert files to financial instruments.
+ * <p>
+ * The main parser is for FpML. See {@link com.opengamma.strata.loader.fpml.FpmlDocumentParser}.
+ */
+package com.opengamma.strata.loader.fpml;
