+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/fx-ex01-fx-spot.xml
