+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/fx-ex03-fx-fwd.xml
