+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/fx-ex07-non-deliverable-forward.xml
