+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/fx-ex08-fx-swap.xml
