+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex01-vanilla-swap.xml
