+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex02-stub-amort-swap.xml
