+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex03-compound-swap.xml
