+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex04-arrears-stepup-fee-swap.xml
