+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex05-long-stub-swap.xml
