+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex06-xccy-swap.xml
