+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex07-ois-swap.xml
