+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex08c-fra-namespace.xml
