+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex25-fxnotional-swap.xml
