+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex26-fxnotional-swap-with-cfs.xml
