+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex28-bullet-payments.xml
