+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex30-swap-comp-avg-relative-date.xml
