+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex32-zero-coupon-swap.xml
