+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/ird-ex35-inverse-floater-inverse-vs-floating.xml
