+++ modules/loader/src/test/resources/com/opengamma/strata/loader/fpml/td-ex01-simple-term-deposit.xml
